Endrer navn på DokumentRad til DokumentListItem og flytter den til en egen fil
