+++ src/app/komponenter/dokumentoversikt/Dokumentoversikt.tsx
@@ -1,17 +1,6 @@
 'use client';
 
-import {
-    Alert,
-    BodyShort,
-    Box,
-    Button,
-    Link,
-    Loader,
-    Pagination,
-    Skeleton,
-    Table,
-    VStack,
-} from '@navikt/ds-react';
+import { Alert, BodyShort, Box, Button, Link, Pagination, Skeleton, Table } from '@navikt/ds-react';
 import {
     TableBody,
     TableDataCell,
@@ -22,13 +11,9 @@ import {
 } from '@navikt/ds-react/Table';
 import { appUrl } from '@/app/util/miljø';
 import React, { useEffect, useState } from 'react';
-import {
-    Datotype,
-    DokumentInfo,
-    Journalpost,
-    Journalposttype,
-} from '@/app/typer/api/Dokumentoversikt';
-import { List, ListItem } from '@navikt/ds-react/List';
+import { Datotype, Journalpost, Journalposttype } from '@/app/typer/api/Dokumentoversikt';
+import { List } from '@navikt/ds-react/List';
+import { DokumentListItem } from './DokumentListItem';
 
 async function hentDokumenter(): Promise<Journalpost[]> {
     const response = await fetch(`${appUrl}/api/dokumenter`);
@@ -143,7 +128,7 @@ export function Dokumentoversikt() {
                                     content={
                                         <List>
                                             {journalpost.dokumenter?.map(dokument => (
-                                                <DokumentRad
+                                                <DokumentListItem
                                                     key={dokument.dokumentInfoId}
                                                     journalpost={journalpost}
                                                     dokument={dokument}
@@ -188,103 +173,6 @@ export function Dokumentoversikt() {
             </>
         );
     }
-}
-
-enum DokumentStatus {
-    IDLE = 'IDLE',
-    IKKE_TILGANG = 'IKKE_TILGANG',
-    LASTER = 'LASTER',
-    FEIL = 'FEIL',
-}
-
-type DokumentRadProps = {
-    journalpost: Journalpost;
-    dokument: DokumentInfo;
-};
-
-async function hentDokument(journalpostId: string, dokumentInfoId: string): Promise<Blob> {
-    const response = await fetch(
-        `${appUrl}/api/dokument?journalpostId=${journalpostId}&dokumentInfoId=${dokumentInfoId}`,
-        { headers: { Accept: 'application/pdf' } }
-    );
-
-    if (!response.ok) {
-        throw new Error(`Failed to fetch document: ${response.statusText}`);
-    }
-
-    return await response.blob();
-}
-
-function DokumentRad({ journalpost, dokument }: DokumentRadProps) {
-    const dokumentTittel = dokument.tittel || 'Dokument uten tittel';
-
-    const harTilgang = dokument.dokumentvarianter.some(variant => variant?.brukerHarTilgang);
-
-    const [dokumentStatus, setDokumentStatus] = useState(
-        harTilgang ? DokumentStatus.IDLE : DokumentStatus.IKKE_TILGANG
-    );
-
-    const hentOgVisDokument = async () => {
-        if (
-            dokumentStatus === DokumentStatus.IKKE_TILGANG ||
-            dokumentStatus === DokumentStatus.LASTER
-        ) {
-            return;
-        }
-
-        setDokumentStatus(DokumentStatus.LASTER);
-
-        try {
-            const blob = await hentDokument(journalpost.journalpostId, dokument.dokumentInfoId);
-            const url = URL.createObjectURL(blob);
-            window.open(url, '_blank');
-            setDokumentStatus(DokumentStatus.IDLE);
-        } catch (error) {
-            console.error('Feil ved henting og visning av dokument:', error);
-            setDokumentStatus(DokumentStatus.FEIL);
-        }
-    };
-
-    return (
-        <ListItem>
-            <VStack>
-                {harTilgang ? (
-                    <>
-                        <span>
-                            <Link
-                                href=""
-                                aria-busy={dokumentStatus === 'LASTER'}
-                                onClick={e => {
-                                    e.preventDefault();
-                                    hentOgVisDokument();
-                                }}
-                            >
-                                {dokumentTittel}
-                            </Link>
-                            {dokumentStatus === 'LASTER' && (
-                                <Box as="span" marginInline="2 0">
-                                    <Loader size="xsmall" title="Laster..." />
-                                </Box>
-                            )}
-                        </span>
-                        {dokumentStatus === 'FEIL' && (
-                            <Alert variant="error" size="small">
-                                Det oppstod en feil under vising av dokumentet. Vennligst prøv igjen
-                                senere.
-                            </Alert>
-                        )}
-                    </>
-                ) : (
-                    <>
-                        <span>{dokumentTittel}</span>
-                        <Alert variant="info" size="small">
-                            Du har ikke tilgang til dette dokumentet.
-                        </Alert>
-                    </>
-                )}
-            </VStack>
-        </ListItem>
-    );
 }
 
 const spørsmålOmDokumenterMangler = (
